+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Add `config.items.write-to-file` feature where a user can directly write the
+  theme content into a specified file, including between optional start and end
+  markers
+
 ## [0.30.0] - 2026-01-14
 
 ### Added